backend/requirements.txt
django == 3.0.3
django-markdownx
gunicorn == 20.0.4
libsass
django-compressor
django-sass-processor
pillow
mysqlclient == 1.4.6
django-slack
django-mdeditor
Markdown
djangorestframework
django-filter
django-cors-headers
djangorestframework-simplejwt
djoser
sendgrid
